backend/requirements.txt
uwsgi
flask
boto3
openai
python-dotenv